+++ pyproject.toml
@@ -1,0 +1,22 @@
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "sortedset"
+version = "0.1.0"
+description = "平衡 Θ(N) 分木の SortedSet です。PyPy で動きます。平衡二分木系より速いと思います。"
+authors = [{name = "tatyam"}]
+readme = "README.md"
+license-files = ["LICENSE"]
+requires-python = ">=3.9"
+
+[project.urls]
+Homepage = "https://github.com/tatyam-prime/SortedSet"
+Repository = "https://github.com/tatyam-prime/SortedSet.git"
+
+[dependency-groups]
+dev = ["pytest>=8.4.2"]
+
+[tool.setuptools.packages.find]
+include = ["sortedset", "sortedset.codon"]
