+++ sortedset/BucketList.py
@@ -1,30 +1,23 @@
-# https://github.com/tatyam-prime/SortedSet/blob/main/codon/BucketList.py
+# https://github.com/tatyam-prime/SortedSet/blob/main/sortedset/BucketList.py
 import math
-from typing import ClassVar, Generator
+from typing import Generic, Iterable, Iterator, TypeVar
+T = TypeVar('T')
 
-class BucketList[T]:
-    size: int
-    a: list[list[T]]
-    BUCKET_RATIO: ClassVar[int] = 16
-    SPLIT_RATIO: ClassVar[int] = 24
-
-    def __init__(self) -> None:
-        self.size = 0
-        self.a = []
-
-    def __init__(self, a: Generator[T]) -> None:
-        self.__init__(list(a))
+class BucketList(Generic[T]):
+    BUCKET_RATIO = 16
+    SPLIT_RATIO = 24
     
-    def __init__(self, a: list[T]) -> None:
+    def __init__(self, a: Iterable[T] = []) -> None:
+        a = list(a)
         n = self.size = len(a)
         num_bucket = int(math.ceil(math.sqrt(n / self.BUCKET_RATIO)))
         self.a = [a[n * i // num_bucket : n * (i + 1) // num_bucket] for i in range(num_bucket)]
 
-    def __iter__(self) -> Generator[T]:
+    def __iter__(self) -> Iterator[T]:
         for i in self.a:
             for j in i: yield j
 
-    def __reversed__(self) -> Generator[T]:
+    def __reversed__(self) -> Iterator[T]:
         for i in reversed(self.a):
             for j in reversed(i): yield j
     
@@ -34,15 +27,9 @@ class BucketList[T]:
             if x != y: return False
         return True
     
-    def __ne__(self, other) -> bool:
-        return not self.__eq__(other)
-
     def __len__(self) -> int:
         return self.size
     
-    def __bool__(self) -> bool:
-        return self.size > 0
-
     def __repr__(self) -> str:
         return "BucketList" + str(self.a)
     
@@ -65,8 +52,7 @@ class BucketList[T]:
     def insert(self, i: int, x: T) -> None:
         "Insert x at the i-th position. / O(√N)"
         if self.size == 0:
-            if i != 0 and i != -1:
-                raise IndexError("index out of range")
+            if i != 0 and i != -1: raise IndexError
             self.a = [[x]]
             self.size = 1
             return
@@ -78,7 +64,7 @@ class BucketList[T]:
             for b, a in enumerate(self.a):
                 if i <= len(a): return self._insert(a, b, i, x)
                 i -= len(a)
-        raise IndexError("index out of range")
+        raise IndexError
 
     def append(self, x: T) -> None:
         "Append x to the end of the list. / amortized O(1)"
@@ -89,7 +75,7 @@ class BucketList[T]:
         a = self.a[-1]
         return self._insert(a, len(self.a) - 1, len(a), x)
     
-    def extend(self, a: Generator[T]) -> None:
+    def extend(self, a: Iterable[T]) -> None:
         for x in a: self.append(x)
     
     def __getitem__(self, i: int) -> T:
@@ -101,8 +87,8 @@ class BucketList[T]:
             for a in self.a:
                 if i < len(a): return a[i]
                 i -= len(a)
-        raise IndexError("index out of range")
-
+        raise IndexError
+    
     def _pop(self, a: list[T], b: int, i: int) -> T:
         ans = a.pop(i)
         self.size -= 1
@@ -119,7 +105,7 @@ class BucketList[T]:
             for b, a in enumerate(self.a):
                 if i < len(a): return self._pop(a, b, i)
                 i -= len(a)
-        raise IndexError("index out of range")
+        raise IndexError
 
     def count(self, x: T) -> int:
         "Return the number of occurrences of x. / O(N)"
@@ -129,8 +115,8 @@ class BucketList[T]:
         "Return the index of the first occurrence of x, raise ValueError if not found. / O(N)"
         for i, y in enumerate(self):
             if x == y: return i
-        raise ValueError("value not in list")
-
+        raise ValueError
+    
     def remove(self, x: T) -> None:
         "Remove the first occurrence of x, raise ValueError if not found. / O(N)"
         self.pop(self.index(x))
@@ -143,5 +129,5 @@ class BucketList[T]:
         self.a.reverse()
         for a in self.a: a.reverse()
 
-    def copy(self):
+    def copy(self) -> 'BucketList[T]':
         return BucketList(self)
