+++ sortedset/SortedSet.py
@@ -1,54 +1,44 @@
-# https://github.com/tatyam-prime/SortedSet/blob/main/codon/SortedMultiset.py
+# https://github.com/tatyam-prime/SortedSet/blob/main/sortedset/SortedSet.py
 import math
 from bisect import bisect_left, bisect_right
-from typing import ClassVar, Generator, Optional
+from typing import Generic, Iterable, Iterator, TypeVar
+T = TypeVar('T')
 
-class SortedMultiset[T]:
-    size: int
-    a: list[list[T]]
-    BUCKET_RATIO: ClassVar[int] = 16
-    SPLIT_RATIO: ClassVar[int] = 24
-
-    def __init__(self) -> None:
-        self.size = 0
-        self.a = []
-
-    def __init__(self, a: Generator[T]) -> None:
-        self.__init__(list(a))
+class SortedSet(Generic[T]):
+    BUCKET_RATIO = 16
+    SPLIT_RATIO = 24
     
-    def __init__(self, a: list[T]) -> None:
-        "Make a new SortedMultiset from a list. / O(N) if sorted / O(N log N)"
-        n = self.size = len(a)
+    def __init__(self, a: Iterable[T] = []) -> None:
+        "Make a new SortedSet from iterable. / O(N) if sorted and unique / O(N log N)"
+        a = list(a)
+        n = len(a)
         if any(a[i] > a[i + 1] for i in range(n - 1)):
             a.sort()
+        if any(a[i] >= a[i + 1] for i in range(n - 1)):
+            a, b = [], a
+            for x in b:
+                if not a or a[-1] != x:
+                    a.append(x)
+        n = self.size = len(a)
         num_bucket = int(math.ceil(math.sqrt(n / self.BUCKET_RATIO)))
         self.a = [a[n * i // num_bucket : n * (i + 1) // num_bucket] for i in range(num_bucket)]
 
-    def __iter__(self) -> Generator[T]:
+    def __iter__(self) -> Iterator[T]:
         for i in self.a:
             for j in i: yield j
 
-    def __reversed__(self) -> Generator[T]:
+    def __reversed__(self) -> Iterator[T]:
         for i in reversed(self.a):
             for j in reversed(i): yield j
     
-    def __eq__(self, other: SortedMultiset[T]) -> bool:
-        if len(self) != len(other): return False
-        for x, y in zip(self, other):
-            if x != y: return False
-        return True
-
-    def __ne__(self, other: SortedMultiset[T]) -> bool:
-        return not self.__eq__(other)
-
+    def __eq__(self, other) -> bool:
+        return list(self) == list(other)
+    
     def __len__(self) -> int:
         return self.size
     
-    def __bool__(self) -> bool:
-        return self.size > 0
-    
     def __repr__(self) -> str:
-        return "SortedMultiset" + str(self.a)
+        return "SortedSet" + str(self.a)
     
     def __str__(self) -> str:
         s = str(list(self))
@@ -65,22 +55,20 @@ class SortedMultiset[T]:
         a, _, i = self._position(x)
         return i != len(a) and a[i] == x
 
-    def count(self, x: T) -> int:
-        "Count the number of x."
-        return self.index_right(x) - self.index(x)
-
-    def add(self, x: T) -> None:
-        "Add an element. / O(√N)"
+    def add(self, x: T) -> bool:
+        "Add an element and return True if added. / O(√N)"
         if self.size == 0:
             self.a = [[x]]
             self.size = 1
-            return
+            return True
         a, b, i = self._position(x)
+        if i != len(a) and a[i] == x: return False
         a.insert(i, x)
         self.size += 1
         if len(a) > len(self.a) * self.SPLIT_RATIO:
             mid = len(a) >> 1
             self.a[b:b+1] = [a[:mid], a[mid:]]
+        return True
     
     def _pop(self, a: list[T], b: int, i: int) -> T:
         ans = a.pop(i)
@@ -95,26 +83,26 @@ class SortedMultiset[T]:
         if i == len(a) or a[i] != x: return False
         self._pop(a, b, i)
         return True
-
-    def lt(self, x: T) -> Optional[T]:
+    
+    def lt(self, x: T) -> T | None:
         "Find the largest element < x, or None if it doesn't exist."
         for a in reversed(self.a):
             if a[0] < x:
                 return a[bisect_left(a, x) - 1]
 
-    def le(self, x: T) -> Optional[T]:
+    def le(self, x: T) -> T | None:
         "Find the largest element <= x, or None if it doesn't exist."
         for a in reversed(self.a):
             if a[0] <= x:
                 return a[bisect_right(a, x) - 1]
 
-    def gt(self, x: T) -> Optional[T]:
+    def gt(self, x: T) -> T | None:
         "Find the smallest element > x, or None if it doesn't exist."
         for a in self.a:
             if a[-1] > x:
                 return a[bisect_right(a, x)]
 
-    def ge(self, x: T) -> Optional[T]:
+    def ge(self, x: T) -> T | None:
         "Find the smallest element >= x, or None if it doesn't exist."
         for a in self.a:
             if a[-1] >= x:
@@ -130,7 +118,7 @@ class SortedMultiset[T]:
             for a in self.a:
                 if i < len(a): return a[i]
                 i -= len(a)
-        raise IndexError("index out of range")
+        raise IndexError
     
     def pop(self, i: int = -1) -> T:
         "Pop and return the i-th element."
@@ -142,8 +130,8 @@ class SortedMultiset[T]:
             for b, a in enumerate(self.a):
                 if i < len(a): return self._pop(a, b, i)
                 i -= len(a)
-        raise IndexError("index out of range")
-
+        raise IndexError
+    
     def index(self, x: T) -> int:
         "Count the number of elements < x."
         ans = 0
