+++ sortedset/codon/BucketList.py
@@ -1,23 +1,30 @@
-# https://github.com/tatyam-prime/SortedSet/blob/main/BucketList.py
+# https://github.com/tatyam-prime/SortedSet/blob/main/sortedset/codon/BucketList.py
 import math
-from typing import Generic, Iterable, Iterator, TypeVar
-T = TypeVar('T')
+from typing import ClassVar, Generator
 
-class BucketList(Generic[T]):
-    BUCKET_RATIO = 16
-    SPLIT_RATIO = 24
+class BucketList[T]:
+    size: int
+    a: list[list[T]]
+    BUCKET_RATIO: ClassVar[int] = 16
+    SPLIT_RATIO: ClassVar[int] = 24
+
+    def __init__(self) -> None:
+        self.size = 0
+        self.a = []
+
+    def __init__(self, a: Generator[T]) -> None:
+        self.__init__(list(a))
     
-    def __init__(self, a: Iterable[T] = []) -> None:
-        a = list(a)
+    def __init__(self, a: list[T]) -> None:
         n = self.size = len(a)
         num_bucket = int(math.ceil(math.sqrt(n / self.BUCKET_RATIO)))
         self.a = [a[n * i // num_bucket : n * (i + 1) // num_bucket] for i in range(num_bucket)]
 
-    def __iter__(self) -> Iterator[T]:
+    def __iter__(self) -> Generator[T]:
         for i in self.a:
             for j in i: yield j
 
-    def __reversed__(self) -> Iterator[T]:
+    def __reversed__(self) -> Generator[T]:
         for i in reversed(self.a):
             for j in reversed(i): yield j
     
@@ -27,9 +34,15 @@ class BucketList(Generic[T]):
             if x != y: return False
         return True
     
+    def __ne__(self, other) -> bool:
+        return not self.__eq__(other)
+
     def __len__(self) -> int:
         return self.size
     
+    def __bool__(self) -> bool:
+        return self.size > 0
+
     def __repr__(self) -> str:
         return "BucketList" + str(self.a)
     
@@ -52,7 +65,8 @@ class BucketList(Generic[T]):
     def insert(self, i: int, x: T) -> None:
         "Insert x at the i-th position. / O(√N)"
         if self.size == 0:
-            if i != 0 and i != -1: raise IndexError
+            if i != 0 and i != -1:
+                raise IndexError("index out of range")
             self.a = [[x]]
             self.size = 1
             return
@@ -64,7 +78,7 @@ class BucketList(Generic[T]):
             for b, a in enumerate(self.a):
                 if i <= len(a): return self._insert(a, b, i, x)
                 i -= len(a)
-        raise IndexError
+        raise IndexError("index out of range")
 
     def append(self, x: T) -> None:
         "Append x to the end of the list. / amortized O(1)"
@@ -75,7 +89,7 @@ class BucketList(Generic[T]):
         a = self.a[-1]
         return self._insert(a, len(self.a) - 1, len(a), x)
     
-    def extend(self, a: Iterable[T]) -> None:
+    def extend(self, a: Generator[T]) -> None:
         for x in a: self.append(x)
     
     def __getitem__(self, i: int) -> T:
@@ -87,8 +101,8 @@ class BucketList(Generic[T]):
             for a in self.a:
                 if i < len(a): return a[i]
                 i -= len(a)
-        raise IndexError
-    
+        raise IndexError("index out of range")
+
     def _pop(self, a: list[T], b: int, i: int) -> T:
         ans = a.pop(i)
         self.size -= 1
@@ -105,7 +119,7 @@ class BucketList(Generic[T]):
             for b, a in enumerate(self.a):
                 if i < len(a): return self._pop(a, b, i)
                 i -= len(a)
-        raise IndexError
+        raise IndexError("index out of range")
 
     def count(self, x: T) -> int:
         "Return the number of occurrences of x. / O(N)"
@@ -115,8 +129,8 @@ class BucketList(Generic[T]):
         "Return the index of the first occurrence of x, raise ValueError if not found. / O(N)"
         for i, y in enumerate(self):
             if x == y: return i
-        raise ValueError
-    
+        raise ValueError("value not in list")
+
     def remove(self, x: T) -> None:
         "Remove the first occurrence of x, raise ValueError if not found. / O(N)"
         self.pop(self.index(x))
@@ -129,5 +143,5 @@ class BucketList(Generic[T]):
         self.a.reverse()
         for a in self.a: a.reverse()
 
-    def copy(self) -> 'BucketList[T]':
+    def copy(self):
         return BucketList(self)
